requirements.txt
pushbullet.py==0.10.0
yahoo-finance==1.2.1
